map contravariant arrays types properly
Type parameter of array type is mapped in a certain way.
If parameter type variance is IN then corresponding Java type is Object.
For Kotlin annotation of array type, an attempt is made to lookup
type parameter of array in Kotlin types
e.g. annotation of Array<Int> now contains jet/Int.

#KT-3771 Fixed