fix main function detection
The main function used to be detected by name "main" and textual form
of the single parameter type "Array<String>". This is now changed to
rely on actually resolved types (both Array and String). Resolving types
relies on scoping information, which is kinda hard to propagate. That is
the main source of complexity of this patch. At 3 of 4 places that needs
main function detection, we can create a ResolveSession, which provides
a ScopeProvider; on the remaining place (the ControlFlowAnalyzer), scopes
can be obtained from BodiesResolveContext.