Fix built-ins reference resolve tests
Built-in sources are contained now in two directories: compiled and
non-compiled (native). kotlin-plugin.jar has them both under 'jet/', but for
tests we manually specify where to look for the former, as it was for the
latter in LightClassUtil.java.

Simplify code in JetReferenceUtil to have almost the same meaning as in
JetPsiReference.multiResolve()