Updated name of intention action to replace a get method call, with an array access expression
Old name : GetCallReplacementIntention
New name : ReplaceGetCallWithArrayAccessIntention

Tests added : functionLiteral.kt, which is meant to test the case in
which a call is made to get where it is defined in place as function
literal. The test in this case should fail, because this situation is
not applicable.

Issues : Cannot run tests. JUnit tests currently not generated properly.