Removed negation of compareTo() result (used for descending sort)
Negating the result of compareTo()/compare() is a questionable or bad programming practice, since if the return value is Integer.MIN_VALUE, negating the return value won't negate the sign of the result. You can achieve the same intended result by reversing the order of the operands rather than by negating the results.