update_dependencies.xml for AS 3.0
jps-standalone should be re-unzipped even if target directory already exists. It may be another version.

build against UAST bundled in AS 3.0

download correct build of Android Studio (KT-20654)

In the latest version of the download page, there's a commented
block with links to downloads of older versions. Use non-greedy
matching to make sure we pick up the first link (the real one),
not the last one (the archive one).