gradle-plugin: Allow 'manual' dependency downloading
This patch adds an ability to download native dependencies
(clang, sysroots etc) during execution of the `downloadKonanCompiler`
task. The feature can by enabled by setting the
`downloadDependencies` property of the `downloadKonanCompiler` task
to true. Targets for these dependencies may be specified as a string
list using `targets` property of the `downloadKonanCompiler` task.