Revert "Set until-build to "173.4301.25" for AS 3.1"
This reverts commit 489665be77d2c0438f98f04a8760b80859c47c1b.

We have to do this, because otherwise tests will fail massively. This is
because there is a bug in testing framework, which causes API version of
plugin to be tested against AS build number (instead of an API version
of IDEA, on which that AS was built upon).

IMPORTANT. That means that "until build" should be corrected manually
for published plugins (e.g. on plugin portal).