Reusing lambda classes instances for stateless lambdas generated during inlining.
When compiler generates a class for a stateless lambda it also generates
code reusing one single lambda instance by storing it in static INSTANCE field.

This is not true for lambda classes re-generated while performing inlining. It
is possible that the original lambda class template captures lambda value which
will be inlined during chained inlining, in some cases lambdas may become
stateless. See for example KT-23168.

This PR tries to solve this by adding singleton-like pattern to the re-generated class.