Change Kotlin updater to use PluginVersionChecker.
The PluginVersionChecker EP is defined by:
https://github.com/JetBrains/intellij-community/pull/797,
used to customize Kotlin plugin advertisement.

This PR changes the updater to pick a PluginVersionChecker
(the one with the highest priority if there are more than one),
to decide the latest Kotlin plugin version to advertise.

The existing implementation is refactored to IdeaPluginVersionChecker,
an implementation of the PluginVersionChecker EP.