Simplify detailed receiver
Use let in order to define detailedReceiver in a single statement

There's functional relationship from context.call.explicitReceiver to detailedReceiver; identity
when the former is a QualifiedReceiver?, transformToReceiveWithSmartCastInfo() otherwise. Using a
let construct exhibits this relationship explicitly, in the anaonymous function. Incidentally,
it removes with explicitReceiver definition which has a limited scope.