Do not report pre-release errors if release LV is used in pre-release compiler
This commit effectively reverts
d386712903e7d583bb74eb13cc00671733bb1d48.

The reason is that when using a release language version, you can only
"see" the subset of a pre-release library which consists of released and
supported features, so reporting an error is not helpful there. Also, it
presents a problem currently when using kotlinc 1.3 (which is
pre-release) with language version 1.2 (stable) against the bundled
stdlib of version 1.3 (pre-release)

 #KT-21267 Declined