Improve SynchronizedLazy performance
Improves initialization performance of synchronized lazy vals, in the scenario where a large numbers of thread are waiting and the initializer is a time consuming operation, by replacing the mutex with an AtomicBoolean and a CountDownLatch. The AB ensures only one thread invokes the initializer, and the latch alllows all threads to proceed at once after initialization is complete. The previous implementation forces each waiting thread to acquire a mutex before proceeding.